+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v0.0.93 (unreleased)
 
 * Add `min_length`/`max_length` support to `list_metadata` for list length validation
+* Add `min_length`/`max_length` support to `str_metadata` for string length validation
 
 ## v0.0.92 (2026-04-03)